requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-taggit==0.22.2
django-taggit-serializer==0.1.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
oauthlib==2.1.0
Pillow==5.1.0
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
rest-framework-cache==0.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
